+++ arc/command/slash.py
@@ -108,7 +108,8 @@ class SlashCommand(CallableCommandBase[ClientT, hikari.api.SlashCommandBuilder])
             id=id,
             options=[option.to_command_option() for option in self.options.values()],
             default_member_permissions=self.default_permissions,
-            is_dm_enabled=self.is_dm_enabled,
+            integration_types=self.integration_types,
+            context_types=self.context_types,
             is_nsfw=self.is_nsfw,
             name_localizations={str(key): value for key, value in self.name_localizations.items()},
             description_localizations={str(key): value for key, value in self.description_localizations.items()},
@@ -263,7 +264,8 @@ class SlashGroup(CommandBase[ClientT, hikari.api.SlashCommandBuilder]):
             id=id,
             options=[subcmd.to_command_option() for subcmd in self.children.values()],
             default_member_permissions=self.default_permissions,
-            is_dm_enabled=self.is_dm_enabled,
+            integration_types=self.integration_types,
+            context_types=self.context_types,
             is_nsfw=self.is_nsfw,
             name_localizations={str(key): value for key, value in self.name_localizations.items()},
             description_localizations={str(key): value for key, value in self.description_localizations.items()},
@@ -522,7 +524,8 @@ class SlashSubGroup(SubCommandBase[ClientT, SlashGroup[ClientT]]):
                 autodefer=self._autodefer,
                 default_permissions=hikari.UNDEFINED,
                 is_nsfw=hikari.UNDEFINED,
-                is_dm_enabled=hikari.UNDEFINED,
+                integration_types=hikari.UNDEFINED,
+                context_types=hikari.UNDEFINED,
             )
         )
 
@@ -675,7 +678,8 @@ class SlashSubCommand(
                 autodefer=self._autodefer,
                 default_permissions=hikari.UNDEFINED,
                 is_nsfw=hikari.UNDEFINED,
-                is_dm_enabled=hikari.UNDEFINED,
+                integration_types=hikari.UNDEFINED,
+                context_types=hikari.UNDEFINED,
             )
         )
 
@@ -739,7 +743,8 @@ def slash_command(
     description: str = "No description provided.",
     *,
     guilds: t.Sequence[hikari.PartialGuild | hikari.Snowflakeish] | hikari.UndefinedType = hikari.UNDEFINED,
-    is_dm_enabled: bool | hikari.UndefinedType = hikari.UNDEFINED,
+    integration_types: t.Sequence[hikari.ApplicationIntegrationType] | hikari.UndefinedType = hikari.UNDEFINED,
+    context_types: t.Sequence[hikari.ApplicationContextType] | hikari.UndefinedType = hikari.UNDEFINED,
     is_nsfw: bool | hikari.UndefinedType = hikari.UNDEFINED,
     autodefer: bool | AutodeferMode | hikari.UndefinedType = hikari.UNDEFINED,
     default_permissions: hikari.Permissions | hikari.UndefinedType = hikari.UNDEFINED,
@@ -756,8 +761,10 @@ def slash_command(
         The description of the command
     guilds : t.Sequence[hikari.PartialGuild | hikari.Snowflakeish] | hikari.UndefinedType
         The guilds this command should be enabled in, if left as undefined, the command is global
-    is_dm_enabled : bool | hikari.UndefinedType
-        If True, the command is usable in direct messages
+    integration_types : t.Sequence[hikari.ApplicationIntegrationType] | hikari.UndefinedType
+        The integration types this command supports the installation of
+    context_types : t.Sequence[hikari.ApplicationContextType] | hikari.UndefinedType
+        The context types this command can be invoked in
     is_nsfw : bool | hikari.UndefinedType
         If True, the command is only usable in NSFW channels
     autodefer : bool | AutodeferMode | hikari.UndefinedType
@@ -804,7 +811,8 @@ def slash_command(
             description_localizations=description_localizations or {},
             default_permissions=default_permissions,
             guilds=guild_ids,
-            is_dm_enabled=is_dm_enabled,
+            integration_types=integration_types,
+            context_types=context_types,
             is_nsfw=is_nsfw,
         )