+++ arc/context/autocomplete.py
@@ -69,6 +69,20 @@ class AutocompleteData(t.Generic[ClientT, ChoiceT]):
         return self.interaction.member
 
     @property
+    def invocation_context(self) -> hikari.ApplicationContextType:
+        """The context in which the interaction was invoked."""
+        return self.interaction.context
+
+    @property
+    def authorizing_integration_owners(self) -> t.Mapping[hikari.ApplicationIntegrationType, hikari.Snowflake]:
+        """Includes details about the authorizing user or server for the installation(s) relevant to the interaction.
+
+        For apps installed to a user, it can be used to tell the difference between the authorizing user
+        and the user that triggered an interaction (like a message component).
+        """
+        return self.interaction.authorizing_integration_owners
+
+    @property
     def user(self) -> hikari.User:
         """The user that triggered the interaction."""
         return self.interaction.user