+++ arc/plugin.py
@@ -33,8 +33,11 @@ class RESTPluginBase(PluginBase[RESTClientT]):
     autodefer : bool | AutodeferMode
         If True, all commands in this plugin will automatically defer if it is taking longer than 2 seconds to respond.
         This can be overridden on a per-command basis.
-    is_dm_enabled : bool | hikari.UndefinedType
-        Whether commands in this plugin are enabled in DMs
+    integration_types : t.Sequence[hikari.ApplicationIntegrationType] | hikari.UndefinedType
+        The integration types that commands will support the installation of
+        This can be overridden on a per-command basis.
+    context_types : t.Sequence[hikari.ApplicationContextType] | hikari.UndefinedType
+        The context types that commands can be invoked in
         This can be overridden on a per-command basis.
     default_permissions : hikari.Permissions | hikari.UndefinedType
         The default permissions for this plugin
@@ -81,8 +84,11 @@ class GatewayPluginBase(PluginBase[GatewayClientT]):
     autodefer : bool | AutodeferMode
         If True, all commands in this plugin will automatically defer if it is taking longer than 2 seconds to respond.
         This can be overridden on a per-command basis.
-    is_dm_enabled : bool | hikari.UndefinedType
-        Whether commands in this plugin are enabled in DMs
+    integration_types : t.Sequence[hikari.ApplicationIntegrationType] | hikari.UndefinedType
+        The integration types that commands will support the installation of
+        This can be overridden on a per-command basis.
+    context_types : t.Sequence[hikari.ApplicationContextType] | hikari.UndefinedType
+        The context types that commands can be invoked in
         This can be overridden on a per-command basis.
     default_permissions : hikari.Permissions | hikari.UndefinedType
         The default permissions for this plugin
@@ -115,7 +121,8 @@ class GatewayPluginBase(PluginBase[GatewayClientT]):
         default_enabled_guilds: t.Sequence[hikari.Snowflakeish | hikari.PartialGuild]
         | hikari.UndefinedType = hikari.UNDEFINED,
         autodefer: bool | AutodeferMode | hikari.UndefinedType = hikari.UNDEFINED,
-        is_dm_enabled: bool | hikari.UndefinedType = hikari.UNDEFINED,
+        integration_types: t.Sequence[hikari.ApplicationIntegrationType] | hikari.UndefinedType = hikari.UNDEFINED,
+        context_types: t.Sequence[hikari.ApplicationContextType] | hikari.UndefinedType = hikari.UNDEFINED,
         default_permissions: hikari.Permissions | hikari.UndefinedType = hikari.UNDEFINED,
         is_nsfw: bool | hikari.UndefinedType = hikari.UNDEFINED,
     ) -> None:
@@ -123,7 +130,8 @@ class GatewayPluginBase(PluginBase[GatewayClientT]):
             name=name,
             default_enabled_guilds=default_enabled_guilds,
             autodefer=autodefer,
-            is_dm_enabled=is_dm_enabled,
+            integration_types=integration_types,
+            context_types=context_types,
             default_permissions=default_permissions,
             is_nsfw=is_nsfw,
         )